Large. Refactor vuestic plugins.
