+++ packages/ui/src/components/va-dropdown/components/VaDropdownContent/_variables.scss
