+++ packages/ui/src/components/va-dropdown/components/VaDropdownContent/index.ts
@@ -1,4 +1,4 @@
-import withConfigTransport from '../../../services/config-transport/withConfigTransport'
+import withConfigTransport from '../../../../services/config-transport/withConfigTransport'
 import VaDropdownContent from './VaDropdownContent.vue'
 
 export default withConfigTransport(VaDropdownContent)
