+++ packages/ui/src/components/va-dropdown/index.ts
@@ -1,5 +1,5 @@
 import withConfigTransport from '../../services/config-transport/withConfigTransport'
 import VaDropdown from './VaDropdown.vue'
-export { default as VaDropdownContent } from './VaDropdownContent'
+export { default as VaDropdownContent } from './components/VaDropdownContent'
 
 export default withConfigTransport(VaDropdown)
