+++ packages/ui/src/components/va-toast/plugin/index.ts
@@ -1,7 +1,7 @@
 import { App } from 'vue'
-import Notification from './index'
+import Notification from '../index'
 
-export default {
+export const VaToastPlugin = {
   install (app: App) {
     app.config.globalProperties.$vaToast = new Notification(app)
   },
