+++ packages/ui/src/vuestic-plugin/vuestic-plugin-without-components.ts
@@ -1,15 +1,13 @@
 import type { GlobalConfig } from '../main'
 import type { App } from 'vue'
-import { DropdownPopperSubplugin, ToastInstall, GlobalConfigPlugin, ColorHelpersPlugin } from './vuestic-plugins'
+import { GlobalConfigPlugin, VaDropdownPlugin, VaToastPlugin, ColorConfigPlugin } from './vuestic-plugins'
 
 export const VuesticPluginsWithoutComponents = {
   install (app: App, vuesticConfig: GlobalConfig): void {
-    app.use(DropdownPopperSubplugin)
-
-    app.use(ToastInstall)
+    app.use(VaDropdownPlugin)
+    app.use(VaToastPlugin)
 
     app.use(GlobalConfigPlugin, vuesticConfig)
-
-    app.use(ColorHelpersPlugin)
+    app.use(ColorConfigPlugin)
   },
 }
