+++ packages/ui/src/vuestic-plugin/vuestic-plugin.ts
@@ -1,17 +1,7 @@
 import { App } from 'vue'
-
-import { installPlatform } from '../components/va-popup/install'
-import DropdownPopperSubplugin
-  from '../components/va-dropdown/dropdown-popover-subplugin'
-import ToastInstall from '../components/va-toast/install'
-
-import ColorHelpersPlugin from '../services/color-config/color-css-variables-updater'
 import { GlobalConfig } from '../services/global-config/global-config'
-import { GlobalConfigPlugin } from '../services/global-config/global-config-plugin'
-
+import { GlobalConfigPlugin, VaDropdownPlugin, VaToastPlugin, ColorConfigPlugin } from './vuestic-plugins'
 import * as vuesticComponentMap from './vuestic-components'
-
-installPlatform()
 
 export const VuesticPlugin = {
   install (app: App, vuesticConfig: GlobalConfig): void {
@@ -19,12 +9,11 @@ export const VuesticPlugin = {
       app.component(name, component)
     })
 
-    app.use(DropdownPopperSubplugin)
-
-    app.use(ToastInstall)
+    app.use(VaDropdownPlugin)
+    app.use(VaToastPlugin)
 
     app.use(GlobalConfigPlugin, vuesticConfig)
 
-    app.use(ColorHelpersPlugin)
+    app.use(ColorConfigPlugin)
   },
 }
