+++ packages/ui/src/vuestic-plugin/vuestic-plugins.ts
@@ -1,5 +1,4 @@
-export { GlobalConfigPlugin } from '../services/global-config/global-config-plugin'
-export { default as ColorHelpersPlugin } from '../services/color-config/color-css-variables-updater'
-export { default as ToastInstall } from '../components/va-toast/install'
-export { default as DropdownPopperSubplugin }
-  from '../components/va-dropdown/dropdown-popover-subplugin'
+export { GlobalConfigPlugin } from '../services/global-config/plugin'
+export { ColorConfigPlugin } from '../services/color-config/plugin'
+export { VaToastPlugin } from '../components/va-toast/plugin'
+export { VaDropdownPlugin } from '../components/va-dropdown/plugin'
